pushing the models file

+++ backend/Idempotency-gateway/app/models.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel, Field
 class PaymentRequest(BaseModel):
     """Incoming payment request body."""
 
-    amount: float = Field(..., gt=0, description="The payment amount (must be positive)")
+    amount: int = Field(..., gt=0, description="The payment amount (must be positive)")
     currency: str = Field(..., min_length=1, description="Currency code, e.g. GHS, USD")
